HDTool: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@39303 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/c/HDTool/setup.c
@@ -18,7 +18,7 @@ struct RDArgs rda = {{args, strlen(args), 0}, 0, 0, 0, NULL, 0};
 		printf
 		(
 			"setup\n\tdevice=%s unit=%ld\n",
-			(STRPTR)myargs[0], *(LONG *)myargs[1]
+			(STRPTR)myargs[0], (long)*(LONG *)myargs[1]
 		);
 		if (myargs[2])
 		{
